Refactor: Change socket.io files setting
